+++ shapiq/explainer/tabular.py
@@ -86,7 +86,7 @@ class TabularExplainer(Explainer):
         random_state: Optional[int] = None,
         **kwargs,
     ) -> None:
-        from shapiq.games.imputer import ConditionalImputer, MarginalImputer
+        from shapiq.games.imputer import BaselineImputer, ConditionalImputer, MarginalImputer
 
         if index not in AVAILABLE_INDICES:
             raise ValueError(f"Invalid index `{index}`. " f"Valid indices are {AVAILABLE_INDICES}.")
@@ -102,7 +102,15 @@ class TabularExplainer(Explainer):
             self._imputer = ConditionalImputer(
                 self.predict, self.data, random_state=random_state, **kwargs
             )
-        elif isinstance(imputer, MarginalImputer) or isinstance(imputer, ConditionalImputer):
+        elif imputer == "baseline":
+            self._imputer = BaselineImputer(
+                self.predict, self.data, random_state=random_state, **kwargs
+            )
+        elif (
+            isinstance(imputer, MarginalImputer)
+            or isinstance(imputer, ConditionalImputer)
+            or isinstance(imputer, BaselineImputer)
+        ):
             self._imputer = imputer
         else:
             raise ValueError(
@@ -143,7 +151,7 @@ class TabularExplainer(Explainer):
         imputer = self._imputer.fit(x)
 
         # explain
-        interaction_values = self._approximator.approximate(budget=budget, game=imputer)
+        interaction_values = self._approximator(budget=budget, game=imputer)
         interaction_values.baseline_value = self.baseline_value
 
         return interaction_values